refactor: extract CI-failure classification out of pr_monitor
``runtime/pr_monitor.py`` was at 1501 lines, one over the first-party
maintainability guardrail, so CI shard 8 kept failing even after ``loop.py``
was brought back under the limit.

Move the pure CI-failure classification block — the transient/code failure
marker tables, the ruff/lint regexes, ``_log_shows_code_failure``, the
``_failure_has_*_evidence`` predicates, ``_ci_failure_identity``,
``_looks_like_transient_ci_failure`` and the candidate-failure partitioning —
into ``pr_monitor_ci_classification``. The block depends only on the
``CheckFailure`` / ``PRStatus`` wire shapes, never on ``MonitorState`` or
``MonitorConfig``, so the new module imports cleanly with no cycle.

Every symbol is re-exported from ``pr_monitor`` with the ``X as X`` form so
existing import sites (including ``_ci_failure_identity`` and
``_log_shows_docker_registry_timeout`` in the tests) keep resolving there. The
moved text is byte-identical and ``__all__`` is untouched, so behavior and the
public API are unchanged. ``pr_monitor.py`` is now 1372 lines.
